Remove duplicated code in test classes
