+++ tests/test_click_commands/CommandBaseTestingClass.py
@@ -11,25 +11,19 @@ You should have received a copy of the GNU General Public License
 along with Tracker. If not, see <http://www.gnu.org/licenses/>.
 """
 
-import unittest
-import os
-
 from click.testing import CliRunner
 
 from src.main import cli, add_subcommands_to_cli
-
-from src.csv.CSVHandler import CSVHandler
-from src.setup_test_values.setup_test_values import SetupTestValues
+from tests.BaseTestingClass import BaseTestingClass
 
 
-class CommandBaseTestingClass(unittest.TestCase):
+class CommandBaseTestingClass(BaseTestingClass):
     def setUp(self) -> None:
+        super(CommandBaseTestingClass, self).setUp()
+
         self.runner = CliRunner()
         self.output = None
         self.exit_code = -1
-
-        self.csv_handler = CSVHandler()
-        self.set_upper = SetupTestValues()
 
         add_subcommands_to_cli()
 
@@ -49,19 +43,3 @@ class CommandBaseTestingClass(unittest.TestCase):
 
     def check_for_exit_code_not_zero(self) -> None:
         self.assertNotEqual(self.exit_code, 0)
-
-# TODO: Duplicated code in this class and CSVBaseTestingClass()
-    def clean_and_init_tracker_file(self) -> None:
-        self.remove_tracker_file()
-        self.csv_handler.init_tracker_csv_file()
-
-    def remove_tracker_file(self) -> None:
-        try:
-            os.remove(self.csv_handler.tracker_file)
-        except FileNotFoundError:
-            # File is not existing, that's good, nothing to do for us.
-            pass
-
-    def setup_test_values(self, number_of_entries: int = 4) -> None:
-        self.set_upper.set_number_of_entries(number_of_entries)
-        self.set_upper.setup()
