+++ tests/test_csv/CSVBaseTestingClass.py
@@ -11,19 +11,14 @@ You should have received a copy of the GNU General Public License
 along with Tracker. If not, see <http://www.gnu.org/licenses/>.
 """
 
-import unittest
-import os
-
 import pandas
 
-from src.csv.CSVHandler import CSVHandler
-from src.setup_test_values.setup_test_values import SetupTestValues
+from tests.BaseTestingClass import BaseTestingClass
 
 
-class CSVBaseTestingClass(unittest.TestCase):
+class CSVBaseTestingClass(BaseTestingClass):
     def setUp(self) -> None:
-        self.csv_handler = CSVHandler()
-        self.set_upper = SetupTestValues()
+        super(CSVBaseTestingClass, self).setUp()
 
     def check_for_correct_column_names(self) -> None:
         self.assertTrue(self.columns_names_are_correct())
@@ -38,17 +33,6 @@ class CSVBaseTestingClass(unittest.TestCase):
         data_frame = pandas.read_csv(self.csv_handler.tracker_file)
         return list(data_frame.columns)
 
-    def clean_and_init_tracker_file(self) -> None:
-        self.remove_tracker_file()
-        self.csv_handler.init_tracker_csv_file()
-
-    def remove_tracker_file(self) -> None:
-        try:
-            os.remove(self.csv_handler.tracker_file)
-        except FileNotFoundError:
-            # File is not existing, that's good, nothing to do for us.
-            pass
-
     def get_contents_of_tracker_file_with_replaced_nans(self) -> pandas.DataFrame:
         data = self.get_contents_of_tracker_file()
         return data.fillna("")
@@ -56,7 +40,3 @@ class CSVBaseTestingClass(unittest.TestCase):
     def get_contents_of_tracker_file(self) -> pandas.DataFrame:
         data_frame = pandas.read_csv(self.csv_handler.tracker_file)
         return data_frame
-
-    def setup_test_values(self, number_of_entries: int = 4) -> None:
-        self.set_upper.set_number_of_entries(number_of_entries)
-        self.set_upper.setup()
